requirements.txt
pip
Pyome
pymongo
dnspython
discord.py
asyncio